Feat: local storage
Created local storage class which can store and retrieve the project
array and the current selected project id. Updated render and projects
modules to use the new local storage methods. Moved the new date function
from the project constructor to the createNewProject methods so as to prevent
new projects getting the date and time of when the class was created.