Feat: update project title heading
On load and when changing project the project title name header updates
with the selected name of the project. The header also persists on reload.
Was pretty simple to implement, apart from the fact I was creating a new
render class in the index file which meant that the project array wasn't
being created so was undefined.