Feat: delete projects
Added button to bottom of task container which deletes the selected
project from the project array, saves the array to the local storage and
then updates the project title to the default. Also fixed a bug where by
when adding a new project it wasn't set as the selected project straight
away.