Feat: add task
Create task and add it to the selected project. Had to figure out why
at first the projects weren't being add to the correct project, was due
to stale references to the local storage. Solved it by passing the project
array and the selected project ID to the add task method.