Fix: edit and update task rather than create new
Fixed issue where when editing a task a new task was created rather than
editing and updating the original. This is also meant that the modal was
dynamically update depending on if it was to add or edit a task. If the
edit button was clicked it would update the title and button in the modal.
Refactored the addTask method so that it included a check whether editingTaskId
existed it would only exist if the edit button had been clicked. This way
it would either add or update the task.